% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a csv file into a dataframe}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the CSV file to be opened.}
}
\value{
A data frame constructed from the contents of the CSV file passed.
}
\description{
This function takes a CSV file name as an argument, and attempts to read the specified
file into a data.frame with the default settings. It gives an error if the file does
not exist.
}
\examples{
\dontrun{farsread(filename = "accident_2013.csv")}
\dontrun{farsread(filename = "accident_2014.csv")}


}
