% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read the month and year of incidents from multiple years of FARS data.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The year(s) of the data of interest in a vector or list.}
}
\value{
This function returns a data frame with two columns, MONTH and year, with a row for each FARS incident
     in the specified years.
}
\description{
This function reads one or more years of FARS data into a data frame, and returns a data frame
containing only the month and the year of each incident. Will give errors for any specified
year that does not have a specified value, and if the dplyr::%>% function has not been imported.
}
\examples{
\dontrun{fars_read_years(2014)}
\dontrun{fars_read_years(c(2014,2015,2016)}

}
