% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Gives the number of FARS incident per month in each specified year.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The year(s) of the data of interest in a vector or list.}
}
\value{
This function returns a data frame with years as columns and month numbers as row numbers.
}
\description{
This function creates a table of the month and year of each incident in the FARS database
for one or more specified years, then computes and returns a summary table of the number of incidents
with the months in rows and years in columns.  Will give an error for any specified year
that does not have an associated data file, as well as an additional error if none
of the specified years are valid.
}
\examples{
\dontrun{fars_summarize_years(2014)}
\dontrun{fars_summarize_years(c(2014,2015,2016)}

}
